Code cleanup done
