+++ frontend/src/Routes/User/UserProfile/index.jsx
@@ -1,57 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from "react-router-dom";
-import { Card, Descriptions, Skeleton } from "antd";
-import PropTypes from 'prop-types';
+import { Card, Skeleton } from "antd";
 
 import { connect, storeType } from '../../../Store';
 import { getProfileInfoByUserId } from '../../../Api/User';
-import { numGradeToStrGrade } from '../../../Helpers/grades';
 import { CenterView } from '../../../Layouts/index';
 import "./index.css";
-
-
-const User = ({ 
-  email, firstName, grades, 
-  languages, lastName, progressLevels }) => {
-
-  const languagesTuple = languages.map((lang, i) => (
-    [lang, grades[i], progressLevels[i]]
-  ));
-  
-  return (
-    <>
-      <Descriptions title="User Info">
-        <Descriptions.Item label="Full Name" className={"capitalize-text"}>
-          {[firstName, lastName].join(' ')}
-        </Descriptions.Item>
-        <Descriptions.Item label="Email">{email}</Descriptions.Item>
-      </Descriptions>
-      { languagesTuple.length ? languagesTuple.map( tuple => 
-        <Descriptions title={tuple[0]}>
-          <Descriptions.Item label="Grade">
-            {numGradeToStrGrade(tuple[1])}
-          </Descriptions.Item>
-          <Descriptions.Item label="Progress Level">
-            {tuple[2]}
-          </Descriptions.Item>
-        </Descriptions>
-      ) : <p>No languages found.</p>}
-    </>
-    
-  );
-}
-
-User.propTypes = {
-  userId: PropTypes.number,
-  firstName: PropTypes.string.isRequired,
-  lastName: PropTypes.string.isRequired,
-  email: PropTypes.string.isRequired,
-  languages: PropTypes.arrayOf(PropTypes.string).isRequired,
-  grades: PropTypes.arrayOf(PropTypes.number).isRequired,
-  progressLevels: PropTypes.arrayOf(PropTypes.number).isRequired,
-};
-
-
+import UserView from '../UserView';
 
 const UserProfile = ({ store: { userId } }) => {
   const params = useParams();
@@ -75,7 +30,7 @@ const UserProfile = ({ store: { userId } }) => {
     <CenterView>
       <Card title="Profile Page" style={styles.card}>
         <Skeleton loading={!profile} title={true} active>
-          <User {...profile} />
+          <UserView {...profile} />
         </Skeleton>
       </Card>
     </CenterView>
@@ -93,4 +48,3 @@ const styles = {
 };
 
 export default connect(UserProfile);
-export { User };
