--- frontend/src/Routes/User/Users/index.css
@@ -1,14 +1,0 @@
-.ant-advanced-search-form {
-    padding: 24px;
-    background: #fbfbfb;
-    border: 1px solid #d9d9d9;
-    border-radius: 6px;
-}
-
-.ant-advanced-search-form .ant-form-item {
-    display: flex;
-}
-
-.ant-advanced-search-form .ant-form-item-control-wrapper {
-    flex: 1;
-}
